Moving all code from Tip1 to a separate class
